requirements.txt
fastapi
langchain
langchain-openai
openai
langchain-community
uvicorn[standard]
python-dotenv
tqdm